+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 
 - I'm doing my best to optimise EmuGUI by adapting the New VM dialog.
 - For that, the plug-in reader has been prepared.
+- The delcache scripts have been updated.
 
 ## Known issues